Snuffles finished

+++ snuffles/Eye.pde
@@ -1,7 +1,7 @@
 class Eye 
 {
-  int x, y;
-  int size;
+  float x, y;
+  float size;
   float angle = 0.0;
   
   Eye(int tx, int ty, int ts) 
@@ -11,7 +11,7 @@ class Eye
     size = ts;
  }
 
-  void update(int mx, int my) 
+  void update(float mx, float my) 
   {
     angle = atan2(my-y, mx-x);
   }
@@ -19,12 +19,12 @@ class Eye
   void display() 
   {
     pushMatrix();
-    translate(x, y);
-    fill(255);
-    ellipse(0, 0, size, size);
-    rotate(angle);
-    fill(153, 204, 0);
-    ellipse(size/4, 0, size/2, size/2);
+      translate(x, y);
+      fill(255);
+      ellipse(0, 0, size, size);
+      rotate(angle);
+      fill(153, 204, 0);
+      ellipse(size/4, 0, size/2, size/2);
     popMatrix();
   }
 }
